Change ScriptEvalUtil logic to be contained in an API class, plus necessary changes for that to work
