+++ data/org.eclipse.birt.data/src/org/eclipse/birt/data/engine/expression/CompiledExpression.java
@@ -12,17 +12,12 @@
 
 package org.eclipse.birt.data.engine.expression;
 
-import org.eclipse.birt.core.script.ScriptContext;
-import org.eclipse.birt.data.engine.core.DataException;
-import org.mozilla.javascript.Scriptable;
+import org.eclipse.birt.data.engine.api.BaseCompiledExpression;
 
 /** 
- * This class encapulates an DtE expression that has been analyzed, rewritten,
- * and compiled during report query preparation. An instance of its derived class
- * is given to the factory as a handle. The factory uses this handle to evaluate
- * the compiled expression at query execution time.
+ * Defines constants to be returned by the inherited getType( ) method
  */
-public abstract class CompiledExpression
+public abstract class CompiledExpression extends BaseCompiledExpression
 {
 	/*
 	 * Constants returned by getType()
@@ -53,16 +48,5 @@ public abstract class CompiledExpression
 	 * An invalid expression
 	 */
 	final static int TYPE_INVALID_EXPR		= 5;
-
-	/**
-	 * gets the type of the compiled expression
-	 */
-	public abstract int getType();
-	
-	/**
-	 * Evaluates this expression
-	 */
-	public abstract Object evaluate( ScriptContext context, Scriptable scope )
-		throws DataException;	
 	
 }
